prints all the list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -15,9 +15,9 @@ size_t print_list(const list_t *h)
 	while (h != NULL)
 	{
 		if (h->str == NULL)
-			printf("[%d] %s\n", 0, "(nil)");
+			printf("[%lu] %s\n", (unsigned long)0, "(nil)");
 		else
-			printf("[%d] %s\n", h->len, h->str);
+			printf("[%lu] %s\n",(unsigned long) h->len, h->str);
 		h = h->next;
 		nodes++;
